// src/yeying/api/common/code.ts
export enum ApplicationCodeEnum {
  /** APPLICATION_CODE_UNKNOWN - 未知编码，是一种占位符 */
  APPLICATION_CODE_UNKNOWN = 0,
  /** APPLICATION_CODE_MARKET - 社区集市，提供应用、服务以及智能合约的发布、上线、下线等服务，用户可以在应用集市组装自己的应用； */
  APPLICATION_CODE_MARKET = 1,
  /** APPLICATION_CODE_ASSET - 资产应用 */
  APPLICATION_CODE_ASSET = 2,
  /** APPLICATION_CODE_KNOWLEDGE - 知识库应用 */
  APPLICATION_CODE_KNOWLEDGE = 3,
  /** APPLICATION_CODE_KEEPER - 智能管家应用，包括钱包 */
  APPLICATION_CODE_KEEPER = 4,
  /** APPLICATION_CODE_SOCIAL - 社交应用 */
  APPLICATION_CODE_SOCIAL = 5,
  /** APPLICATION_CODE_WORKBENCH - 工作台应用 */
  APPLICATION_CODE_WORKBENCH = 6,
  UNRECOGNIZED = -1,
}
// ...
export enum ServiceCodeEnum {
  /** SERVICE_CODE_UNKNOWN - 未知编码，是一种占位符 */
  SERVICE_CODE_UNKNOWN = 0,
  /** SERVICE_CODE_NODE - 网络节点供应商 */
  SERVICE_CODE_NODE = 1,
  /** SERVICE_CODE_WAREHOUSE - 仓储服务供应商 */
  SERVICE_CODE_WAREHOUSE = 2,
  /** SERVICE_CODE_AGENT - 智能体供应商 */
  SERVICE_CODE_AGENT = 3,
  /** SERVICE_CODE_MCP - 模型上下文供应商 */
  SERVICE_CODE_MCP = 4,
  UNRECOGNIZED = -1,
}
// ...
export enum ApplicationStatusEnum {
  /** APPLICATION_STATUS_UNKNOWN - 未知编码，是一种占位符 */
  APPLICATION_STATUS_UNKNOWN = 0,
  /** APPLICATION_STATUS_CREATED - 已创建 */
  APPLICATION_STATUS_CREATED = 1,
  /** APPLICATION_STATUS_AUDITED - 已审计 */
  APPLICATION_STATUS_AUDITED = 2,
  /** APPLICATION_STATUS_REFUSED - 已拒绝 */
  APPLICATION_STATUS_REFUSED = 3,
  /** APPLICATION_STATUS_OFFLINE - 已下架 */
  APPLICATION_STATUS_OFFLINE = 4,
  /** APPLICATION_STATUS_ONLINE - 已上线 */
  APPLICATION_STATUS_ONLINE = 5,
  UNRECOGNIZED = -1,
}