// src/yeying/api/common/model.ts
export interface ApplicationMetadata {
  /** 应用所有者 */
  owner: string;
  /** 身份所在网络 */
  network: string;
  /** 身份地址 */
  address: string;
  /** 应用身份 */
  did: string;
  /** 应用版本 */
  version: number;
  /** 应用哈希 */
  hash: string;
  /** 应用名称 */
  name: string;
  /** 应用编号 */
  code: ApplicationCodeEnum;
  /** 应用描述 */
  description: string;
  /** 访问路径 */
  location: string;
  /** 应用依赖的服务编码，用,隔开 */
  serviceCodes: ServiceCodeEnum[];
  /** 应用头像 */
  avatar: string;
  /** 应用创建时间 */
  createdAt: string;
  /** 应用更新时间 */
  updatedAt: string;
  /** 签名 */
  signature: string;
  /** 代码包下载路径 */
  codePackagePath: string;
}