// src/yeying/api/application/application.ts
export interface SearchApplicationResponse {
  header: MessageHeader | undefined;
  body: SearchApplicationResponseBody | undefined;
}

export interface SearchApplicationResponseBody {
  status: ResponseStatus | undefined;
  applications: ApplicationMetadata[];
  page: ResponsePage | undefined;
}

export interface CreateApplicationRequest {
  header: MessageHeader | undefined;
  body: CreateApplicationRequestBody | undefined;
}

export interface CreateApplicationRequestBody {
  application: ApplicationMetadata | undefined;
}

export interface CreateApplicationResponse {
  header: MessageHeader | undefined;
  body: CreateApplicationResponseBody | undefined;
}

export interface CreateApplicationResponseBody {
  status: ResponseStatus | undefined;
  application: ApplicationMetadata | undefined;
}

export interface ApplicationDetailRequest {
  header: MessageHeader | undefined;
  body: ApplicationDetailRequestBody | undefined;
}

export interface ApplicationDetailRequestBody {
  did: string;
  version: number;
}

export interface ApplicationDetailResponse {
  header: MessageHeader | undefined;
  body: ApplicationDetailResponseBody | undefined;
}

export interface ApplicationDetailResponseBody {
  status: ResponseStatus | undefined;
  application: ApplicationMetadata | undefined;
}

export interface DeleteApplicationRequest {
  header: MessageHeader | undefined;
  body: DeleteApplicationRequestBody | undefined;
}

export interface DeleteApplicationRequestBody {
  did: string;
  version: number;
}

export interface DeleteApplicationResponse {
  header: MessageHeader | undefined;
  body: DeleteApplicationResponseBody | undefined;
}

export interface DeleteApplicationResponseBody {
  status: ResponseStatus | undefined;
}
// ...
export const CreateApplicationResponse: MessageFns<CreateApplicationResponse> = {
  encode(message: CreateApplicationResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      CreateApplicationResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateApplicationResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateApplicationResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = CreateApplicationResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateApplicationResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? CreateApplicationResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: CreateApplicationResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = CreateApplicationResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateApplicationResponse>, I>>(base?: I): CreateApplicationResponse {
    return CreateApplicationResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateApplicationResponse>, I>>(object: I): CreateApplicationResponse {
    const message = createBaseCreateApplicationResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? CreateApplicationResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseCreateApplicationResponseBody(): CreateApplicationResponseBody {
  return { status: undefined, application: undefined };
}

export const CreateApplicationResponseBody: MessageFns<CreateApplicationResponseBody> = {
  encode(message: CreateApplicationResponseBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      ResponseStatus.encode(message.status, writer.uint32(10).fork()).join();
    }
    if (message.application !== undefined) {
      ApplicationMetadata.encode(message.application, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateApplicationResponseBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateApplicationResponseBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = ResponseStatus.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.application = ApplicationMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateApplicationResponseBody {
    return {
      status: isSet(object.status) ? ResponseStatus.fromJSON(object.status) : undefined,
      application: isSet(object.application) ? ApplicationMetadata.fromJSON(object.application) : undefined,
    };
  },

  toJSON(message: CreateApplicationResponseBody): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = ResponseStatus.toJSON(message.status);
    }
    if (message.application !== undefined) {
      obj.application = ApplicationMetadata.toJSON(message.application);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateApplicationResponseBody>, I>>(base?: I): CreateApplicationResponseBody {
    return CreateApplicationResponseBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateApplicationResponseBody>, I>>(
    object: I,
  ): CreateApplicationResponseBody {
    const message = createBaseCreateApplicationResponseBody();
    message.status = (object.status !== undefined && object.status !== null)
      ? ResponseStatus.fromPartial(object.status)
      : undefined;
    message.application = (object.application !== undefined && object.application !== null)
      ? ApplicationMetadata.fromPartial(object.application)
      : undefined;
    return message;
  },
};
// ...
export const ApplicationDetailResponse: MessageFns<ApplicationDetailResponse> = {
  encode(message: ApplicationDetailResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      ApplicationDetailResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationDetailResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationDetailResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = ApplicationDetailResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationDetailResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? ApplicationDetailResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: ApplicationDetailResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = ApplicationDetailResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationDetailResponse>, I>>(base?: I): ApplicationDetailResponse {
    return ApplicationDetailResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationDetailResponse>, I>>(object: I): ApplicationDetailResponse {
    const message = createBaseApplicationDetailResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? ApplicationDetailResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseApplicationDetailResponseBody(): ApplicationDetailResponseBody {
  return { status: undefined, application: undefined };
}

export const ApplicationDetailResponseBody: MessageFns<ApplicationDetailResponseBody> = {
  encode(message: ApplicationDetailResponseBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      ResponseStatus.encode(message.status, writer.uint32(10).fork()).join();
    }
    if (message.application !== undefined) {
      ApplicationMetadata.encode(message.application, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationDetailResponseBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationDetailResponseBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = ResponseStatus.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.application = ApplicationMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationDetailResponseBody {
    return {
      status: isSet(object.status) ? ResponseStatus.fromJSON(object.status) : undefined,
      application: isSet(object.application) ? ApplicationMetadata.fromJSON(object.application) : undefined,
    };
  },

  toJSON(message: ApplicationDetailResponseBody): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = ResponseStatus.toJSON(message.status);
    }
    if (message.application !== undefined) {
      obj.application = ApplicationMetadata.toJSON(message.application);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationDetailResponseBody>, I>>(base?: I): ApplicationDetailResponseBody {
    return ApplicationDetailResponseBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationDetailResponseBody>, I>>(
    object: I,
  ): ApplicationDetailResponseBody {
    const message = createBaseApplicationDetailResponseBody();
    message.status = (object.status !== undefined && object.status !== null)
      ? ResponseStatus.fromPartial(object.status)
      : undefined;
    message.application = (object.application !== undefined && object.application !== null)
      ? ApplicationMetadata.fromPartial(object.application)
      : undefined;
    return message;
  },
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };